Add original version of script from September 2016
This is the first version of the script which I wrote in September 2016.
It was designed to create an iso file from the InstallESD.dmg file in an
Install OS X app for OS X 10.9 and later. It did not accommodate the
different structure of InstallESD.dmg on 10.7 and 10.8. It did not
address the name change to macOS in 10.12. It did not handle the changed
location of BaseSystem.dmg on 10.13 and later.

SQ UPDATE COPYRIGHT LINE