Add new version of script from March 2020
This is a new version of the script from March 2020 rewritten from
scratch.

Rather than creating the iso manually, it uses the createinstallmedia
command line program embedded within the Install app for OS X 10.9 and
later, so it still doesn't work for 10.7 or 10.8. It handles the fact
that createinstallmedia requires different arguments on different OS X
versions and it tries to work around a bug in the Install macOS Sierra
12.6.06 app. (https://feedbackassistant.apple.com/feedback/7628029)

It creates the iso using hdiutil makehybrid because that automatically
creates an iso of the smallest size and having no free space, whereas
with hdiutil convert we had to compute the size and resize the source
disk image first, which sometimes failed when Spotlight wrote some data
to the disk after we computed the space used but before we resized it.
hdiutil makehybrid also results in a read-only iso, which is what we
want. hdiutil convert created a writable iso presumably since the source
disk image was writable.