Add new version of script from June 19, 2020
This is a new version of the script rewritten from scratch again over
the course of a few days leading up to June 19, 2020. This is a work in
progress and not a finished script.

I wasn't happy with how createinstallmedia required different args on
different OS X versions, didn't work without a hack in the rereleased
macOS Sierra installer, and didn't exist on 10.7 and 10.8. I've also
seen reports that createinstallmedia from newer OS installers won't run
on older OS versions, and I would like to be able to create an installer
of any OS version on any OS version. So this version of the script goes
back to constructing the disk image from which the iso is made manually.
This also allows it to work with just an InstallESD.dmg file, without
having the surrounding Install app.

It works with Install apps from 10.7 through 10.12. For 10.9 and later,
it constructs the temporary disk image by restoring the base system
image and then copying the base system dmg and packages to it from the
InstallESD. For 10.7 and 10.8 it just restores the InstallESD to the
temporary image. It doesn't yet handle the changed location of the base
system dmg on 10.13 and later.

A goal with this version of the script is to have an accurate progress
bar for the entire installation process. This is not complete yet.