Add changes from June 20, 2020
Convert script from bash to zsh. Bash doesn't have a good way for
subshells to change global variables, and this was starting to be a
problem with implementing the progress bar code.

When creating an iso for 10.7 or 10.8, don't copy the InstallESD.dmg to
the temporary image; use a shadow file instead. This is faster and more
efficient.