Support 10.13 and newer; support converting pkg
Now supports creating an iso from a macOS 10.13 or later installer,
where the BaseSystem.dmg isn't on the install ESD but rather is in the
same directory as the InstallESD.dmg.

Also now supports creating an iso from a dmg that contains either an
install app or an installer pkg that installs an install app.